+++ src/dev-stream-client.js
@@ -167,3 +167,6 @@ DevStreamClient._streams = {};
 
 module.exports = { DevStreamClient };
 
+
+
+