+++ src/index.d.ts
@@ -212,3 +212,6 @@ export interface EndpointOptions {
 export function createLexiaApp(options?: AppOptions): any;
 export function addStandardEndpoints(app: any, options?: EndpointOptions): any;
 
+
+
+